refactor: Update import paths for Debrief feature types to new directory structure
