+++ libs/shared-types/src/index.ts
@@ -47,21 +47,21 @@ export type {
   DebriefPointFeature,
   DebriefAnnotationFeature,
   DebriefFeatureCollection
-} from './types/FeatureCollection';
+} from './types/features/debrief_feature_collection';
 
 // Re-export state types
-export type { TimeState } from './types/TimeState';
-export type { ViewportState } from './types/ViewportState';
-export type { SelectionState } from './types/SelectionState';
-export type { EditorState } from './types/EditorState';
-export type { CurrentState } from './types/CurrentState';
+export type { TimeState } from './types/states/time_state';
+export type { ViewportState } from './types/states/viewport_state';
+export type { SelectionState } from './types/states/selection_state';
+export type { EditorState } from './types/states/editor_state';
+export type { CurrentState } from './types/states/current_state';
 
 // Import types for union
 import type {
   DebriefTrackFeature,
   DebriefPointFeature,
   DebriefAnnotationFeature
-} from './types/FeatureCollection';
+} from './types/features/debrief_feature_collection';
 
 // Union type for any Debrief feature
 export type DebriefFeature = DebriefTrackFeature | DebriefPointFeature | DebriefAnnotationFeature;
